Refactor with a vengeance
Rename comm to controller, because honestly the name 'comm' confuses me.

Start splitting the bifrost package up - eventually it'll have a bit for
the message protocol, a bit for the core command set, etc etc.

Start making parsers and emitters for the various core command requests
and responses - these'll slowly be moved out of the controller package
with the idea that they can be re-used outside of baps3d.

Other various frippery.